requirements.txt
discord.py
pymongo[srv]
python-dotenv
Flask
beautifulsoup4
requests
Pillow
